// apps/web/src/api/generated/FormControlDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "../../../../../crates/nexus-api/bindings/serde_json/JsonValue";
import type { SchemaHintDto } from "./SchemaHintDto";

export type FormControlDto = { control_id: string, kind: string, label: string, help_text: string | null, mode: string, default_value: JsonValue | null, widget_hint: string | null, schema_hint: SchemaHintDto | null, };